requirements.txt
python-telegram-bot==20.7
requests==2.31.0
psutil==5.9.8